server/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
psycopg2-binary
sqlalchemy
alembic
passlib[bcrypt]
pyjwt
stripe
email-validator
python-multipart
